+++ src/scope/repositories/sources.js
@@ -274,8 +274,6 @@ to quickly fix the issue, please delete the object at "${this.objects().objectPa
     flattenedCompilerDependencies,
     flattenedTesterDependencies,
     message,
-    exactVersion,
-    releaseType,
     specsResults
   }: {
     source: ConsumerComponent,
@@ -285,8 +283,6 @@ to quickly fix the issue, please delete the object at "${this.objects().objectPa
     flattenedCompilerDependencies: BitIds,
     flattenedTesterDependencies: BitIds,
     message: string,
-    exactVersion: ?string,
-    releaseType: string,
     specsResults?: any
   }): Promise<ModelComponent> {
     const objectRepo = this.objects();
@@ -313,7 +309,7 @@ to quickly fix the issue, please delete the object at "${this.objects().objectPa
       flattenedTesterDependencies,
       specsResults
     });
-    component.addVersion(version, releaseType, exactVersion);
+    component.addVersion(version, source.version);
     objectRepo.add(version).add(component);
 
     files.forEach(file => objectRepo.add(file.file));
@@ -328,7 +324,7 @@ to quickly fix the issue, please delete the object at "${this.objects().objectPa
     component: ModelComponent,
     version: Version,
     message: string,
-    releaseType: string = DEFAULT_BIT_RELEASE_TYPE
+    versionToAdd: string
   ): Promise<ModelComponent> {
     const [username, email] = await Promise.all([
       globalConfig.get(CFG_USER_NAME_KEY),
@@ -340,7 +336,7 @@ to quickly fix the issue, please delete the object at "${this.objects().objectPa
       email,
       date: Date.now().toString()
     };
-    component.addVersion(version, releaseType);
+    component.addVersion(version, versionToAdd);
     return this.put({ component, objects: [version] });
   }